+++ include/netdev.h
@@ -148,7 +148,7 @@ int fecmxc_register_mii_postcall(struct eth_device *dev, int (*cb)(int));
 /* Implements the Micrel SMI-protocol for writing to extended register set
  * of Micrel switches (e.g. KSZ8863).
  */
-int fecmxc_smi_write(struct eth_device *dev, uint8_t regAddr, uint16_t data);
+int fecmxc_smi_write(struct eth_device *dev, uint8_t regAddr, uint8_t data);
 #endif
 
 #endif /* _NETDEV_H_ */
